demo-server/requirements.txt
flask
fastapi
uvicorn
pandas
scikit-learn
xgboost
joblib
python-dotenv
pydantic
google-generativeai
